Partial implementation of the add command. Configuration + filesystem linking
There's still some things that need to be sorted. This is so far
untested on Windows. Travis and the Windows-based CI could be an option.

Configuring is done with SCons' Automake-like API. This is still a WIP,
and is managed in the upstream LunarWatcher/SConsStandard repo. Caching
probably needs to be explicitly added, whenever I figure out how.

The major change in this commit is the addition of
FilesystemUtil.hpp/-.cpp. FilesystemUtil is in charge of expanding user
paths (~/, ~user/) into absolute paths. ~user is not supported on
Windows due to various implementation details that require more testing
on Windows than I can currently do. Anyway, it should work fine on
UNIX-based systems, but I haven't tested it enough to say for sure. It
at least works on Linux, from what I can tell.

Some database support has been added. Reading and writing is ready, but
the AddCommand additionally needs IDs and UIDs. These are details I'll
revisit when I finish the command.