[Broken - see description] Fixes to how the ID is displayed; added project syntax.
This is a partial commit that doesn't fix everything I wanted to fix. This commit is made so I can attempt to switch to a different table library. As much as I like tabulate, there's now various bugs related to it I can't work around easily. In an attempt to work around this, I'm trying different libraries, which unfortunately means I need to push an incomplete commit, so I can revert as little as possible if I fuck up