Adds workaround for consumers for #17; disable CI
Because of the issues presented in #17, I've both disabled the CI, and
added a workaround for anyone who still wants to build the program. This
workaround won't be added to the CI, because it can be removed as soon
as Conan 1.28.0 comes out (doesn't take long), and  because it'll
increase CI processing time further (and it's already bad enough,
especially on Windows).

The workaround is by cloning
https://github.com/memsharded/conan/tree/fix/scons_system_libs,
and specifically that branch, and placing it somewhere useful.

If you're using a virtualenv, activate the one you use for building
Taskranger before following the instructions here:
https://docs.conan.io/en/latest/installation.html#install-from-source

TL;DR: clone the repo linked in this message, install requirements, then
finally:

    export CUSTOM_CONAN=/path/to/where/you/cloned/conan
    scons

Again, this is only necessary until Conan 1.28.0 is released