Start the tabulate -> printable migration
This commit moves TermUtils and EnvVars (the hell was I thinking when I named that? Anyway) to printable. Printable is a new table printing library (get it?) that I'm now making to replace tabulate. Tabulate does a lot of stuff in a rather stupid way, and the easiest way to fix it at this point is to roll my own. This kinda sorta makes #21 unfixable by moving the problem to a different repo, but it kinda doesn't matter.

This shouldn't cause any problems with the CI, but I'm pushing and committing to make sure there isn't a problem *cough cough Windows cough*