Remove travis...
This isn't optimal, but Travis has announced it's switching to a
different pricing model, which unfortunately seems to affect
open-source.

For one, this project would lose its Mac build, and I'd like to maintain
OS compatibility. Secondly, the entire CI might stop working if, as the
pricing page suggested, the credits are given once and are never
"reloaded". The free plan is aimed as a trial plan, which isn't really
what I'm looking for.

Unfortunately, that only leaves one option: GH Actions. I'm keeping
Windoze on AppVeyor because I really like being able to see the status
of more builds than one if one fails, which GH Actions is pretty shit
at.